requirements.txt
python-dotenv
InquirerPy
beautifulsoup4
requests
colorama
termcolor
tqdm
pytest
pytest-timeout
pyinstaller
